+++ test/e2e/advance/report_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package e2e
+package advance
 
 import (
 	"context"
@@ -23,7 +23,6 @@ import (
 
 	istio_mixer_v1 "istio.io/api/mixer/v1"
 	pb "istio.io/api/mixer/v1/config/descriptor"
-	"istio.io/mixer/pkg/adapter"
 	"istio.io/mixer/pkg/template"
 	"istio.io/mixer/test/e2e"
 	spyAdapter "istio.io/mixer/test/e2e/advance/spyAdapter"
@@ -192,17 +191,4 @@ func TestReport(t *testing.T) {
 
 		tt.validate(t, err, spyAdapters)
 	}
-}
-
-// ConstructAdapterInfos constructs spyAdapters for each of the adptBehavior. It returns
-// the constructed spyAdapters along with the adapters Info functions.
-func ConstructAdapterInfos(adptBehaviors []spyAdapter.AdapterBehavior) ([]adapter.InfoFn, []*spyAdapter.Adapter) {
-	var adapterInfos []adapter.InfoFn = make([]adapter.InfoFn, 0)
-	spyAdapters := make([]*spyAdapter.Adapter, 0)
-	for _, b := range adptBehaviors {
-		sa := spyAdapter.NewSpyAdapter(b)
-		spyAdapters = append(spyAdapters, sa)
-		adapterInfos = append(adapterInfos, sa.GetAdptInfoFn())
-	}
-	return adapterInfos, spyAdapters
 }
